third_party/blink/web_tests/http/tests/media/media-playback-while-not-visible-permission-policy-pause-when-hidden-cross-origin.html

<!DOCTYPE html>
<title>Test behavior of HTMLMediaElements with the media-playback-while-not-rendered permission policy</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/media-playback-while-not-visible-utils.js"></script>
<body>
<script>
  internals.settings.setAutoplayPolicy('no-user-gesture-required');

  async function setUpIframeMediaPlayerInitialState(t) {
    if (document.readyState !=='complete'){
      await new Promise(resolve => {
        window.addEventListener('load', resolve)
      })
    }

    let iframe = document.createElement('iframe');
    iframe.id = 'video-iframe';
    iframe.allow = "media-playback-while-not-visible 'none'; autoplay *"
    iframe.src = "http://localhost:8000/media/resources/media-playback-while-not-visible-iframe.html";
    document.body.appendChild(iframe);
    await new Promise(resolve => {
      iframe.addEventListener('load', resolve)
    })

    let initial_status = await queryPlayerStatus(iframe);
    if (initial_status !== 'Paused') {
      await pauseMediaInIframe(iframe);
    }
    assert_equals(await queryPlayerStatus(iframe), 'Paused');

    t.add_cleanup(() => {
      document.body.removeChild(iframe);
    });
  }

  function hideFrame(iframe, type) {
    if (type === 'display') {
      iframe.style.setProperty('display', 'none');
    } else if (type === 'visibility') {
      iframe.style.setProperty('visibility', 'hidden');
    } else if (type === 'zeroSize') {
      iframe.style.setProperty('width', '0');
      iframe.style.setProperty('height', '0');
    }
  }

  function showFrame(iframe, type) {
    if (type === 'display') {
      iframe.style.setProperty('display', 'block');
    } else if (type === 'visibility') {
      iframe.style.setProperty('visibility', 'visible');
    } else if (type === 'zeroSize') {
      iframe.style.removeProperty('width');
      iframe.style.removeProperty('height');
    }
  }

  let visibility_types = ['display', 'visibility', 'zeroSize'];
  for (let type of visibility_types) {
    promise_test(async t => {
      await setUpIframeMediaPlayerInitialState(t);
      let iframe = document.getElementById('video-iframe');

      // Start playing media in the iframe.
      assert_equals(await playMediaInIframe(iframe), 'Success');
      assert_equals(await queryPlayerStatus(iframe), 'Playing');

      // Hiding the iframe should pause the embedded video element playback.
      let statechange_promise = expectMediaPlayerStateChangeInIframe(iframe);
      hideFrame(iframe, type);
      assert_equals(await statechange_promise, 'Pause');

      // Attempting to play while hidden should fail.
      assert_equals(await playMediaInIframe(iframe), 'NotAllowedError');
      assert_equals(await queryPlayerStatus(iframe), 'Paused');

      // When the iframe is shown again, playing should be allowed. Give it some
      // time for the cross-origin iframe visibility state to propagate.
      showFrame(iframe, type);
      return new Promise (resolve => setTimeout(async () => {
        statechange_promise = expectMediaPlayerStateChangeInIframe(iframe);
        assert_equals(await playMediaInIframe(iframe), 'Success');
        assert_equals(await statechange_promise, 'Play');
        assert_equals(await queryPlayerStatus(iframe), 'Playing');
        resolve();
      }, 2000));
    }, "Media playback should pause for hidden iframes when permission policy is set (type = " + type + ")");
  }
</script>
</body>
